refactor: move handlers for reusability
